ppc: Minor tweaks to get the arch compile again (WIP)
- Factor in types changes (introduction of intptr_t)
- Align JamFiles syntax with in progress architectures (arm/sparc)
- Xorriso doesn't support much of the mkisofs options (anymore ?)
- (After a correct bootstrap) one should be able to build @minimum-raw and haiku-boot-cd again
Reviewed-on: https://review.haiku-os.org/c/haiku/+/1983

+++ src/add-ons/kernel/bus_managers/pci/arch/ppc/openfirmware/pci_openfirmware.cpp
@@ -30,7 +30,7 @@ status_t
 ppc_openfirmware_pci_controller_init(void)
 {
 	char path[256];
-	int cookie = 0;
+	intptr_t cookie = 0;
 	while (of_get_next_device(&cookie, 0, "pci", path, sizeof(path))
 			== B_OK) {
 dprintf("ppc_openfirmware_pci_controller_init(): pci device node: %s\n", path);
